requirements.txt
# 核心NLP库
spacy>=3.6.0
scikit-learn>=1.3.0
nltk>=3.8.0

# Web框架
streamlit>=1.28.0

# 数据可视化
plotly>=5.15.0

# 文本处理
python-docx>=0.8.11
PyPDF2>=3.0.1
pandas>=2.0.0
numpy>=1.24.0

# 开发工具
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
pre-commit>=3.3.0

# 部署
gunicorn>=21.2.0 